LayoutTests/platform/ios-sim-deprecated/fast/events/check-defocus-event-order-when-triggered-by-tab-expected.txt

This test checks that when a user focuses on the "first input" text field, modifies its contents, and then defocuses it by pressing the tab key that the following DOM events are fired in order: Focus, Change, Blur, Focus (on "second input"). Note, this test must be run by Dump Render Tree.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


  
PASS fired event is (firstInput, focus).
FAIL fired event is (firstInput, blur). Should be (firstInput, change).
FAIL fired event is (secondInput, focus). Should be (firstInput, blur).
FAIL should have fired event (secondInput, focus). But did not.

TEST COMPLETE
